fix: extracted Routes component with early return
Rather than duplicating SideContent in each return from the main App.js
function, this keeps all logic for rendering SideContent in one spot and
instead moves the logic for what else to render into the Routes
component. This component is responsible for checking if entries is
empty and returning early with the empty card if so.

Other cleanup:

* remove commented-out code
* use `function` fore all components
* put `SideContent` before `App` – not sure about this one
* re-retrieve `entries` from `appState` rather than passing as prop